requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.5
Django==3.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
dnspython==2.3.0
filelock==3.7.1
idna==3.4
imutils==0.5.4
numpy==1.24.1
opencv-python==4.7.0.68
Pillow==9.4.0
platformdirs==2.5.2
PyJWT==2.6.0
pymongo==4.3.3
pytz==2022.6
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.16.2
